[共通] update ImageProcessing

+++ Siv3D/include/Siv3D/AdaptiveThresholdMethod.hpp
@@ -14,7 +14,7 @@
 
 namespace s3d
 {
-	enum class AdaptiveThresholdMethod : uint32
+	enum class AdaptiveThresholdMethod : uint8
 	{
 		Mean,
